# backend/swagger.yaml
openapi: 3.1.0
info:
  title: CoinCraft API
  description: |
    Financial literacy learning platform API for children, teens, parents, and teachers. 
    Each endpoint is mapped to specific user stories to ensure alignment with application requirements.
  version: 1.0.0
paths:
...
  /api/users/{user_id}/conversion-requests:
    get:
      tags:
      - Redemptions
      summary: Get Conversion Requests
      description: Get money conversion (redemption) requests for a user.
      operationId: get_conversion_requests_api_users__user_id__conversion_requests_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: status
        in: query
        required: false
        schema:
          type: string
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RedemptionRequestRead'
                title: Response Get Conversion Requests Api Users  User Id  Conversion
                  Requests Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Redemptions
      summary: Create Conversion Request
      description: Create a new request to convert virtual coins to real money.
      operationId: create_conversion_request_api_users__user_id__conversion_requests_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedemptionRequestCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Conversion Request Api Users  User Id  Conversion
                  Requests Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/parents/{parent_id}/redemption-requests:
    get:
      tags:
      - Redemptions
      summary: Get Parent Redemption Requests
      description: Get redemption requests from a parent's children.
      operationId: get_parent_redemption_requests_api_parents__parent_id__redemption_requests_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: parent_id
        in: path
        required: true
        schema:
          type: string
          title: Parent Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RedemptionRequestRead'
                title: Response Get Parent Redemption Requests Api Parents  Parent
                  Id  Redemption Requests Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/redemption-requests/{request_id}/approve:
    put:
      tags:
      - Redemptions
      summary: Approve Redemption Request
      description: Approve a redemption request.
      operationId: approve_redemption_request_api_redemption_requests__request_id__approve_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionRequestRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/redemption-requests/{request_id}/reject:
    put:
      tags:
      - Redemptions
      summary: Reject Redemption Request
      description: Reject a redemption request.
      operationId: reject_redemption_request_api_redemption_requests__request_id__reject_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedemptionRequestRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RedemptionRequestCreate:
      properties:
        coins_amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Coins Amount
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - coins_amount
      title: RedemptionRequestCreate
    RedemptionRequestRead:
      properties:
        coins_amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Coins Amount
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        cash_amount:
          type: number
          title: Cash Amount
        status:
          type: string
          title: Status
          default: pending
        approved_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Approved By
        approved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Approved At
        created_at:
          type: string
          format: date-time
          title: Created At
        child:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Child
      type: object
      required:
      - coins_amount
      - id
      - user_id
      - cash_amount
      - created_at
      title: RedemptionRequestRead
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login